+++ src/compile/source.rs
@@ -46,7 +46,7 @@ impl Lines {
     }
 
     /// Returns the line and column location of `byte`
-    pub fn location(&self, byte: BytePos, file_id: usize) -> Option<Location> {
+    pub fn location(&self, byte: BytePos, file_id: u64) -> Option<Location> {
         if byte.to_usize() <= self.end {
             let line_index = self.line_number_at_byte(byte);
